Faeture back amazons3

See merge request s05-webmobile2-sub2/S05P12C105!49

+++ backend/src/main/java/com/web/curation/files/PhotoDto.java
@@ -12,7 +12,7 @@ import javax.persistence.*;
 @Table(name = "Photo")
 public class PhotoDto {
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long photoId;
+    private Integer photoId;
 
     @Column(name = "original_filename")
     private String originalName;
